+++ src/Method.php
@@ -38,6 +38,7 @@ class Method
     protected $real_name;
     protected $return = null;
     protected $root;
+    protected $classAliases;
 
     /**
      * @param \ReflectionMethod|\ReflectionFunctionAbstract $method
@@ -45,11 +46,13 @@ class Method
      * @param \ReflectionClass $class
      * @param string|null $methodName
      * @param array $interfaces
+     * @param array $classAliases
      */
-    public function __construct($method, $alias, $class, $methodName = null, $interfaces = [])
+    public function __construct($method, $alias, $class, $methodName = null, $interfaces = [], array $classAliases = [])
     {
         $this->method = $method;
         $this->interfaces = $interfaces;
+        $this->classAliases = $classAliases;
         $this->name = $methodName ?: $method->name;
         $this->real_name = $method->isClosure() ? $this->name : $method->name;
         $this->initClassDefinedProperties($method, $class);
@@ -80,7 +83,7 @@ class Method
      */
     protected function initPhpDoc($method)
     {
-        $this->phpdoc = new DocBlock($method, new Context($this->namespace));
+        $this->phpdoc = new DocBlock($method, new Context($this->namespace, $this->classAliases));
     }
 
     /**
@@ -363,7 +366,7 @@ class Method
         }
         if ($method) {
             $namespace = $method->getDeclaringClass()->getNamespaceName();
-            $phpdoc = new DocBlock($method, new Context($namespace));
+            $phpdoc = new DocBlock($method, new Context($namespace, $this->classAliases));
 
             if (strpos($phpdoc->getText(), '{@inheritdoc}') !== false) {
                 //Not at the end yet, try another parent/interface..