+++ tests/Console/ModelsCommand/Factories/__snapshots__/Test__test__1.php
@@ -4,22 +4,20 @@ declare(strict_types=1);
 
 namespace Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Factories\Models;
 
-use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Factories\Factories\SimpleFactory;
+use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Factories\CustomSpace\ModelWithCustomNamespaceFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 /**
- * Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Factories\Models\Simple
+ * Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Factories\Models\ModelWithCustomNamespace
  *
- * @property integer $id
- * @method static \Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Factories\Factories\SimpleFactory factory(...$parameters)
- * @method static \Illuminate\Database\Eloquent\Builder|Simple newModelQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|Simple newQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|Simple query()
- * @method static \Illuminate\Database\Eloquent\Builder|Simple whereId($value)
+ * @method static \Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Factories\CustomSpace\ModelWithCustomNamespaceFactory factory(...$parameters)
+ * @method static \Illuminate\Database\Eloquent\Builder|ModelWithCustomNamespace newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|ModelWithCustomNamespace newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|ModelWithCustomNamespace query()
  * @mixin \Eloquent
  */
-class Simple extends Model
+class ModelWithCustomNamespace extends Model
 {
     use HasFactory;
 
@@ -30,6 +28,67 @@ class Simple extends Model
      */
     protected static function newFactory()
     {
-        return SimpleFactory::new();
+        return ModelWithCustomNamespaceFactory::new();
     }
+}
+<?php
+
+declare(strict_types=1);
+
+namespace Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Factories\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+/**
+ * Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Factories\Models\ModelWithFactory
+ *
+ * @method static \Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Factories\Factories\ModelWithFactoryFactory factory(...$parameters)
+ * @method static \Illuminate\Database\Eloquent\Builder|ModelWithFactory newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|ModelWithFactory newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|ModelWithFactory query()
+ * @mixin \Eloquent
+ */
+class ModelWithFactory extends Model
+{
+    use HasFactory;
+}
+<?php
+
+declare(strict_types=1);
+
+namespace Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Factories\Models;
+
+/**
+ * Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Factories\Models\ModelWithNestedFactory
+ *
+ * @method static \Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Factories\Factories\ModelWithNestedFactoryFactory factory(...$parameters)
+ * @method static \Illuminate\Database\Eloquent\Builder|ModelWithNestedFactory newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|ModelWithNestedFactory newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|ModelWithNestedFactory query()
+ * @mixin \Eloquent
+ */
+class ModelWithNestedFactory extends ModelWithFactory
+{
+}
+<?php
+
+declare(strict_types=1);
+
+namespace Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Factories\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+/**
+ * Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Factories\Models\ModelWithoutFactory
+ *
+ * @method static \Illuminate\Database\Eloquent\Builder|ModelWithoutFactory newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|ModelWithoutFactory newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|ModelWithoutFactory query()
+ * @mixin \Eloquent
+ */
+class ModelWithoutFactory extends Model
+{
+    use HasFactory;
 }